pyproject.toml
[project]
name = "contro"
version = "0.7.0"
description = "A Controversy-Oriented Model of Dialectical Perspectives"
authors = [
    {name = "Francesca Massarenti",email = "146438454+frammenti@users.noreply.github.com"},
    {name = "Alberto Ciarrocca",email = "99617159+vattelalberto@users.noreply.github.com"}
]
readme = "README.md"
license = "GPL-3.0-or-later"
requires-python = ">=3.12,<4.0.0"
dependencies = [
    "mkdocs-material >=9.6.9",
    "rdflib >=7.1.3",
    "saxonche (>=12.5.0,<12.8.0)",
    "mkdocs-open-in-new-tab >=1.0.8",
    "mkdocs-exclude-search >=0.6.6",
    "mkdocs-bibtex >=4.2.3",
    "mkdocs-minify-plugin >=0.8.0",
    "mkdocs-git-revision-date-localized-plugin >=1.4.5",
    "mkdocs-git-committers-plugin-2 >=2.5.0",
    "markdown-span-attr >=0.3.0",
    "lxml (>=6.0.0,<7.0.0)",
]

[tool.poetry]
package-mode = false
